Added Open Reading Frames.py
I think this is working except there is one reading frame in the example output that I can not figure out. I don't see how the example reading frame is even a possibility. That being said, this code should work assuming that one of the reading frames given on the website was a mistake.